NLP.GenI.Configuration

Documentation

module NLP.GenI.Control

getBuilderType :: Params -> BuilderType

getRanking :: Params -> OtRanking

mainBuilderTypes :: [BuilderType]

emptyParams :: Params

The default parameters configuration

defineParams :: [Flag] -> Params -> Params

treatArgs :: [OptDescr Flag] -> [String] -> IO Params

treatArgsWithParams :: [OptDescr Flag] -> [String] -> Params -> IO Params

usage

Arguments

:: [OptSection]
options

-> String
prog name

-> String

Print out a GenI-style usage message with options divided into sections

basicSections :: [OptSection]

optionsSections :: [OptSection]

processInstructions :: Params -> IO Params

Update the internal instructions list, test suite and case according to the contents of an instructions file.

Basic approach

we always have instructions: if no instructions file, is specified we infer virtual instructions from the test suite flag * the testsuite and testcase flags are focusing tools, they pick out a subset from the instructions

optionsForStandardGenI :: [OptDescr Flag]

Uses the GetOpt library to process the command line arguments. Note that we divide them into basic and advanced usage.

optionsForBasicStuff :: [OptDescr Flag]

optionsForOptimisation :: [OptDescr Flag]

optionsForMorphology :: [OptDescr Flag]

optionsForInputFiles :: [OptDescr Flag]

optionsForBuilder :: [OptDescr Flag]

optionsForTesting :: [OptDescr Flag]

helpOption :: OptDescr Flag

verboseOption :: OptDescr Flag

lexiconOption, macrosOption :: OptDescr Flag

nubBySwitches :: [OptDescr a] -> [OptDescr a]

noArg :: forall f. (Eq f, Typeable f) => (() -> f) -> ArgDescr Flag

reqArg

Arguments

:: forall f x . (Eq f, Typeable f, Eq x, Typeable x)

=> (x -> f)
flag

-> (String -> x)
string reader for flag (probably |id| if already a String)

-> String
description

-> ArgDescr Flag

optArg

Arguments

:: forall f x . (Eq f, Typeable f, Eq x, Typeable x)

=> (x -> f)
flag

-> x
default value

-> (String -> x)
string reader (as in reqArg)

-> String
description

-> ArgDescr Flag

parseFlagWithParsec :: String -> Parser b -> Text -> b

TODO: This is a horrible and abusive use of error

configration files

readGlobalConfig :: IO (Maybe YamlLight)

setLoggers :: YamlLight -> IO ()

module System.Console.GetOpt

module NLP.GenI.Flag

class Typeable a

The class Typeable allows a concrete representation of a type to be calculated.