NLP.GenI.Console

Description

The console user interface including batch processing on entire test suites.

Documentation

consoleGeni :: ProgStateRef -> CustomSem sem -> IO ()

loadNextSuite :: ProgStateRef -> CustomSem sem -> (FilePath, Maybe [Text]) -> IO [TestCase sem]

Used in processing instructions files. Each instruction consists of a suite file and a list of test case names from that file

See http://projects.haskell.org/GenI/manual/command-line.html for how testsuite, testcase, and instructions are expected to interact

(Exported for use by regression testing code)