NLP.GenI.LexicalSelection.Types

Documentation

data PathEqLhs

Left hand side of a path equation

Constructors

PeqInterface Text

PeqJust NodePathEqLhs

PeqUnknown Text

Path equations

data NodePathEqLhs

Path equations can either hit a feature or a node's lexeme attribute

Constructors

PeqFeat Text TopBottom Text

PeqLex Text

data TopBottom

Constructors

Top

Bottom

type PathEqPair = (NodePathEqLhs, GeniVal)

parsePathEq :: Text -> Writer [LexCombineError] PathEqLhs

Parse a path equation using the GenI conventions This always succeeds, but can return Just warning if anything anomalous comes up FIXME : make more efficient

showPathEqLhs :: PathEqLhs -> Text

Warnings

data LexCombineError

Constructors

BoringError Text

FamilyNotFoundError Text

SchemaError [Text] LexCombineError2

data LexCombineError2

Constructors

EnrichError PathEqLhs

StringError Text

showLexCombineError :: LexCombineError -> (Text, Text)

compressLexCombineErrors :: [LexCombineError] -> [LexCombineError]