NLP.GenI.Parser

Test suites

geniTestSuite :: Parser [TestCase SemInput]

geniSemanticInput :: Parser (Sem, Flist GeniVal, [LitConstr])

geniTestSuiteString :: Parser [Text]

Just the String representations of the semantics in the test suite

geniDerivations :: Parser [TestCaseOutput]

This is only used by the script genimakesuite

Trees

geniMacros :: Parser [SchemaTree]

geniTagElems :: Parser [TagElem]

This makes it possible to read anchored trees, which may be useful for debugging purposes.

FIXME: note that this is very rudimentary; we do not set id numbers, parse polarities. You'll have to call some of our helper functions if you want that functionality.

Lexicon and morph

geniLexicon :: Parser [LexEntry]

geniMorphInfo :: Parser [(Text, Flist GeniVal)]

Basics

geniFeats :: GeniValLike v => Parser (Flist v)

geniSemantics :: Parser Sem

geniValue :: Parser GeniVal

geniWords :: Parser Text

Helpers

geniWord :: Parser Text

geniLanguageDef :: GenLanguageDef Text () Identity

tillEof :: Parser a -> Parser a

parseFromFile :: Parser a -> SourceName -> IO (Either ParseError a)

module Text.Parsec

module Text.Parsec.Text