NLP.GenI.Polarity.Types

Documentation

data PolarityKey

Constructors

PolarityKeyAv Text Text

PolarityKeyStr Text

PolarityKeyVar Text
attribute

type SemPols = [Int]

data PolarityAttr

PolarityAttr is something you want to perform detect polarities on.

Constructors

SimplePolarityAttr

Fields

spkAtt :: Text

RestrictedPolarityAttr
RestrictedPolarityKey c att is a polarity key in which we only pay attention to nodes that have the category c. This makes it possible to have polarities for a just a small subset of nodes

Fields

_rpkCat :: Text

rpkAtt :: Text

readPolarityAttrs :: String -> Set PolarityAttr

showPolarityAttrs :: Set PolarityAttr -> String