NLP.GenI.Pretty

Description

This is not a proper pretty printer. I aim is to replace this with a (de-facto) standard library if one should appear

Documentation

class Pretty a where

An alternative Show instance (the idea being that we should reserve Show for outputting actual Haskell)

Minimal implementation is pretty or prettyStr

Methods

pretty :: a -> Text

prettyStr :: a -> String

(<>) :: Monoid m => m -> m -> m

An infix synonym for mappend.

(<+>) :: Text -> Text -> Text

Separated by space unless one of them is empty (in which case just the non-empty one)

above :: Text -> Text -> Text

I think I want ($+$) here but I'm not sure I understand the documentation from the pretty package.

t1 above t2 separates the two by a newline, unless one of them is empty. The vertical equivalent to '(+)'

between :: Text -> Text -> Text -> Text

between l r t == l  t  r

parens :: Text -> Text

parens t puts t between parentheses (())

squares :: Text -> Text

squares t puts t between square brackets ([])

squeezed :: Int -> [Text] -> Text

Puts list items on the same line if they are smaller than a certain width otherwise, puts a newline in between them

prettyCount :: (a -> Text) -> Text -> (a, Int) -> Text

 prettyCount toBlah ""     (x,1) == "blah"
 prettyCount toBlah "foos" (x,1) == "blah"
 prettyCount toBlah ""     (x,4) == "blah ×4"
 prettyCount toBlah "foos" (x,4) == "blah ×4 foos"