NLP.GenI.Simple.SimpleBuilder

Documentation

type Agenda = [SimpleItem]

type AuxAgenda = [SimpleItem]

type Chart = [SimpleItem]

data SimpleStatus

type SimpleState a = BuilderState SimpleStatus a

data SimpleItem

Constructors

SimpleItem

Fields

siId :: ChartId

siSubstnodes :: [NodeName]

siAdjnodes :: [NodeName]

siSemantics :: BitVector

siPolpaths :: BitVector

siNodes :: [GNode GeniVal]

actually a set

siDerived :: Tree Text

siRoot_ :: NodeName

siFoot_ :: Maybe NodeName

siPendingTb :: [NodeName]

siDerivation :: TagDerivation

siGuiStuff :: SimpleGuiItem

simpleBuilder_2p, simpleBuilder_1p :: SimpleBuilder

simpleBuilder :: Bool -> SimpleBuilder

theAgenda :: SimpleStatus -> Agenda

theHoldingPen :: SimpleStatus -> AuxAgenda

theChart :: SimpleStatus -> Chart

theResults :: SimpleStatus -> [SimpleItem]

initSimpleBuilder :: Bool -> Input -> [Flag] -> (SimpleStatus, Statistics)

Creates an initial SimpleStatus.

addToAgenda :: SimpleItem -> SimpleState ()

addToChart :: SimpleItem -> SimpleState ()

genconfig :: SimpleStatus -> [Flag]

data SimpleGuiItem

Things whose only use is within the graphical debugger

Constructors

SimpleGuiItem

Fields

siHighlight :: [Text]

nodes to highlight if there are things wrong with this item, what?

siDiagnostic :: [String]

siFullSem :: Sem

siIdname :: Text

theTrash :: SimpleStatus -> Trash

step :: SimpleStatus -> GenerationPhase

unpackResult :: SimpleItem -> [Output]

Aliases to non-exported functions

testCanAdjoin :: SimpleItem -> TagSite -> Maybe (TagSite, TagSite, Subst)

testIapplyAdjNode :: Bool -> SimpleItem -> SimpleItem -> Maybe SimpleItem

testEmptySimpleGuiItem :: SimpleGuiItem