NLP.GenI.TreeSchema

Description

This module provides basic datatypes specific to Tree Adjoining Grammar tree schemata.

Documentation

type Macros = [SchemaTree]

type SchemaTree = Ttree (GNode SchemaVal)

data Ttree a

Constructors

TT

Fields

params :: [GeniVal]

pfamily :: Text

pidname :: Text

pinterface :: Flist GeniVal

ptype :: Ptype

psemantics :: Maybe Sem

ptrace :: [Text]

tree :: Tree a

data Ptype

Constructors

Initial

Auxiliar

root :: Tree a -> a

rootUpd :: Tree a -> a -> Tree a

foot :: Tree (GNode a) -> GNode a

setLexeme :: [Text] -> Tree (GNode a) -> Tree (GNode a)

Given a lexical item l and a tree node n (actually a subtree with no children), return the same node with the lexical item as its unique child. The idea is that it converts terminal lexeme nodes into preterminal nodes where the actual terminal is the given lexical item

setAnchor :: FullList Text -> Tree (GNode a) -> Tree (GNode a)

Given a lexical item s and a Tree GNode t, returns the tree t' where l has been assigned to the anchor node in t'

lexemeAttributes :: [Text]

Attributes recognised as lexemes, in order of preference

crushTreeGNode :: Tree (GNode SchemaVal) -> Maybe (Tree (GNode GeniVal))

data GNode gv

A single node of a TAG tree.

Constructors

GN

Fields

gnname :: NodeName

gup :: Flist gv

top feature structure

gdown :: Flist gv

bottom feature structure

ganchor :: Bool

False for na nodes

glexeme :: [Text]

[] for na nodes

gtype :: GType

gaconstr :: Bool

gorigin :: Text

for TAG, this would be the elementary tree that this node originally came from

gnnameIs :: NodeName -> GNode gv -> Bool

type NodeName = Text

data GType

Constructors

Subs

Foot

Lex

Other

gCategory :: Flist GeniVal -> Maybe GeniVal

Return the value of the cat attribute, if available

showLexeme :: [Text] -> Text

Fancy disjunction

data SchemaVal

A schema value is a disjunction of GenI values. It allows us to express “fancy” disjunctions in tree schemata, ie. disjunctions over variables and not just atoms (?X;?Y).

Our rule is that that when a tree schema is instantiated, any fancy disjunctions must be “crushed” into a single GeniVal lest it be rejected (see crushOne)

Note that this is still not recursive; we don't have disjunction over schema values, nor can schema values refer to schema values. It just allows us to express the idea that in tree schemata, you can have either variable ?X or ?Y.

crushGNode :: GNode SchemaVal -> Maybe (GNode GeniVal)