NLP.GenI.Warning

Description

Typed warnings as an easier alternative to strings.

This makes it easier to recognise repeated warnings and print them out in a reasonable way

Collection of warnings

data GeniWarnings

This exists because we want the Monoid instance, providing a GenI-specific notion of appending which merges instances of the same error

fromGeniWarnings :: GeniWarnings -> [GeniWarning]

mkGeniWarnings :: [GeniWarning] -> GeniWarnings

sortWarnings :: GeniWarnings -> GeniWarnings

Individual warnings

data GeniWarning

Constructors

LexWarning [LexEntry] LexWarning
A warning that should be repeated for each lexical entry affected

CustomLexWarning Text
A single custom warning

NoLexSelection [Literal GeniVal]
Literals which did not receive any lexical selection

MorphWarning [Text]
Warnings from the morphological realiser

data LexWarning

Constructors

LexCombineAllSchemataFailed

LexCombineOneSchemaFailed LexCombineError

MissingCoanchors Text Int

showGeniWarning :: GeniWarning -> [Text]

A warning may be displayed over several lines