NLP.GenI.Warning.Internal

Documentation

newtype GeniWarnings

This exists because we want the Monoid instance, providing a GenI-specific notion of appending which merges instances of the same error

Constructors

GeniWarnings

Fields

fromGeniWarnings :: [GeniWarning]

mkGeniWarnings :: [GeniWarning] -> GeniWarnings

data GeniWarning

Constructors

LexWarning [LexEntry] LexWarning
A warning that should be repeated for each lexical entry affected

CustomLexWarning Text
A single custom warning

NoLexSelection [Literal GeniVal]
Literals which did not receive any lexical selection

MorphWarning [Text]
Warnings from the morphological realiser

data LexWarning

Constructors

LexCombineAllSchemataFailed

LexCombineOneSchemaFailed LexCombineError

MissingCoanchors Text Int

posort :: Poset a => [a] -> [a]

Sort, treating non-comporable items as equal

sortWarnings :: GeniWarnings -> GeniWarnings

appendWarning :: GeniWarning -> [GeniWarning] -> [GeniWarning]

mergeWarning :: GeniWarning -> GeniWarning -> Maybe GeniWarning

showGeniWarning :: GeniWarning -> [Text]

A warning may be displayed over several lines

type WordFamilyCount = Map (FullList Text, Text) Int

toWfCount :: [LexEntry] -> WordFamilyCount